requirements.txt
streamlit
cohere
python-dotenv